projects/ibkr-options-bot/requirements.txt
ib-insync>=0.9.84
pandas>=2.0
pyyaml>=6.0
requests>=2.31
python-dateutil>=2.8
aiofiles>=23.0
